Fixing credits issues

+++ src/app/lib/validators.punishments.ts
@@ -2,12 +2,12 @@ import { z } from 'zod';
 
 export const punishmentCreateSchema = z.object({
     title: z.string().trim().min(2).max(160),
-    marksDeduction: z.coerce.number().int().min(0).optional(),
+    marksDeduction: z.coerce.number().min(0).optional(),
 });
 
 export const punishmentUpdateSchema = z.object({
     title: z.string().trim().min(2).max(160).optional(),
-    marksDeduction: z.coerce.number().int().min(0).optional(),
+    marksDeduction: z.coerce.number().min(0).optional(),
 }).refine(v => Object.keys(v).length > 0, { message: 'Provide at least one field to update' });
 
 export const punishmentListQuerySchema = z.object({
